Moved all script structs, typedefs and defines to a separate header file, which is included in all other script header files. This way we avoid some header nesting.
