+++ CMakeLists.txt
@@ -46,7 +46,7 @@ add_executable(jabs main.c sample.c sample.h brick.c ion.c ion.h simulation.c si
         options.c options.h spectrum.c spectrum.h fit.c fit.h rotate.c rotate.h detector.c detector.h jabs.c jabs.h
         roughness.c roughness.h script.c script.h generic.c generic.h message.c message.h aperture.c aperture.h
         geostragg.c geostragg.h script_command.c script_command.h script_session.c script_session.h script_file.c
-        script_file.h calibration.c calibration.h)
+        script_file.h calibration.c calibration.h script_generic.h)
 
 target_include_directories(jabs PRIVATE ${GETOPT_INCLUDE_DIR})
