+++ script_session.h
@@ -17,21 +17,8 @@
 
 #include <jibal.h>
 #include "simulation.h"
-#include "script_file.h"
 #include "defaults.h"
-
-#include "script_command.h"
-
-typedef struct script_session {
-    jibal *jibal;
-    struct fit_data *fit;
-    char *output_filename; /* File name for automatic spectra saving */ /* TODO: multidetector! */
-    clock_t start, end;
-    script_file *files[SCRIPT_NESTED_MAX];
-    size_t file_depth;
-    struct script_command *commands;
-    size_t i_det_active;
-} script_session;
+#include "script_generic.h"
 
 script_session *script_session_init(jibal *jibal, simulation *sim); /* sim can be NULL or a previously initialized sim can be given. Note that it will be free'd by script_session_free()! */
 void script_session_free(script_session *s);
